+++ .gitignore
@@ -2,3 +2,4 @@ node_modules/
 frontend/node_modules/
 .env
 .DS_Store
+.claude
