+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "biokeeper-frontend",
+  "name": "@biokeeper/frontend",
   "private": true,
   "version": "0.0.0",
   "type": "module",
