+++ package.json
@@ -1,19 +1,20 @@
 {
-  "name": "biokeeper-dashboard",
+  "name": "biokeeper",
   "version": "1.0.0",
   "description": "ESP32 vegetable maintenance dashboard with UV control, sensors, and a modern React UI.",
-  "main": "server.js",
+  "private": true,
+  "workspaces": [
+    "backend",
+    "frontend"
+  ],
   "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js"
-  },
-  "dependencies": {
-    "cors": "^2.8.5",
-    "dotenv": "^16.3.1",
-    "express": "^4.18.2",
-    "mysql2": "^3.5.1"
+    "dev": "concurrently \"npm run dev:backend\" \"npm run dev:frontend\"",
+    "dev:backend": "npm run dev --workspace=@biokeeper/backend",
+    "dev:frontend": "npm run dev --workspace=@biokeeper/frontend",
+    "start": "npm run start --workspace=@biokeeper/backend",
+    "build": "npm run build --workspace=@biokeeper/frontend"
   },
   "devDependencies": {
-    "nodemon": "^3.0.1"
+    "concurrently": "^8.2.2"
   }
 }
